backend/requirements.txt
# ── Core ────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
django-environ==0.11.2
django-cors-headers==4.3.1

# ── Auth ──────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ── Banco de dados ────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# ── Cache ─────────────────────────────────────────
redis==5.0.4
django-redis==5.4.0

# ── Async Tasks ───────────────────────────────────
celery==5.4.0

# ── Documentação da API ───────────────────────────
drf-spectacular==0.27.2

# ── Dev / Qualidade ───────────────────────────────
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
black==24.4.2
flake8==7.0.0
isort==5.13.2
mypy==1.10.0
django-stubs==5.0.2
